refactor(test): change from individual tests to data driven tests
